Refactored changes
