+++ games/Player.java
@@ -11,7 +11,7 @@ class Player {
   private int faceValueCount = 0;
   private int aceCount = 0;
 
-  public void initialize(Deck deck) {
+  public void dealInitialCards(Deck deck) {
     System.out.println("Initializing Player: ");
     Card card1 = deck.randomCard();
     Card card2 = deck.randomCard();
@@ -21,8 +21,8 @@ class Player {
     if(card2.face == Card.Face.A){
       aceCount++;
     }
-    faceValueCount += Card.getFaceValue(card1);
-    faceValueCount += Card.getFaceValue(card2);
+    faceValueCount += BlackJackUtil.getFaceValue(card1);
+    faceValueCount += BlackJackUtil.getFaceValue(card2);
     System.out.println(card1);
     System.out.println(card2);
   }
@@ -33,8 +33,8 @@ class Player {
       aceCount++;
     }
     System.out.println(card);
-    faceValueCount += Card.getFaceValue(card);
-    faceValueCount = PlayUtil.checkWinLooseCondition(faceValueCount, aceCount);
+    faceValueCount += BlackJackUtil.getFaceValue(card);
+    faceValueCount = BlackJackUtil.checkWinLooseCondition(faceValueCount, aceCount);
   }
 
   public int getCurrentCount(){
